+++ src/test/java/com/kt/bit/csm/blds/PerfomanceTest.java
@@ -94,12 +94,12 @@ public class PerfomanceTest extends AbstractBenchmark {
             if (pv instanceof String) {
                 assertEquals(pv, rs.getString(i));
             } else if (pv instanceof Integer) {
-                assertEquals(((Integer) pv).intValue(), rs.getInt(i));
+                assertEquals(pv, rs.getInt(i));
             } else if (pv instanceof Date) {
                 Date a = rs.getDate(i);
                 assertEquals(pv, a);
             } else if (pv instanceof Long) {
-                assertEquals(((Long) pv).longValue(), rs.getLong(i));
+                assertEquals(pv, rs.getLong(i));
             } else if (pv instanceof byte[]) {
                 assertEquals(pv, rs.getBytes(i));
             } else if (pv instanceof Timestamp) {
@@ -126,11 +126,11 @@ public class PerfomanceTest extends AbstractBenchmark {
             if (pv instanceof String) {
                 assertEquals(pv, rs.getString(i));
             } else if (pv instanceof Integer) {
-                assertEquals(((Integer) pv).intValue(), rs.getInt(i));
+                assertEquals(pv, rs.getInt(i));
             } else if (pv instanceof Date) {
                 assertEquals(pv, rs.getDate(i));
             } else if (pv instanceof Long) {
-                assertEquals(((Long) pv).longValue(), rs.getLong(i));
+                assertEquals(pv, rs.getLong(i));
             } else if (pv instanceof byte[]) {
                 assertEquals(pv, rs.getBytes(i));
             } else if (pv instanceof Timestamp) {
